requirements.txt
Django==5.2.12
djangorestframework==3.17.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.1
django-htmx==1.27.0
Pillow==12.2.0
markdown==3.10.2
bleach==6.3.0
python-dotenv==1.2.2
gunicorn==23.0.0
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.1.0
